msgraph/pyproject.toml
[tool.poetry]
name = "msgraph"
version = "0.1.0"
description = "Microsoft 365 Graph API Search Provider"
authors = ["Scott Mountenay <scott@lightsonsoftware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
connexion = { version = "2.14.2", extras = ["swagger-ui"] }
flask = "2.2.5"
python = "^3.11"
pydantic = "2.8.2"
azure-core = "1.23.1"
azure-identity = "1.10.0"
python-dotenv = "1.0.0"
msgraph-core = "0.2.2"
mypy = "^1.4.1"
nltk = "^3.8.1"
requests = "^2.31.0"
gunicorn = "^22.0.0"
aiohttp = "^3.9.4"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
